COG-542: TestCognitoServiceFacade is reworked to stand-alone class
